chore: moar refactor/doc
